fix(webhook): restore shared auth validation across features
